change data type to float

+++ src/main/java/sprt/Algorithm.java
@@ -14,7 +14,7 @@ public class Algorithm {
 		return XTXInverseXT.mmul(Y);
 	}
 
-	public static double compute_cBetaHat(Matrix c, Matrix betaHat) {
+	public static float compute_cBetaHat(Matrix c, Matrix betaHat) {
 		return c.mmul(betaHat).get(0);
 	}
 
@@ -28,11 +28,11 @@ public class Algorithm {
 		return null;
 	}
 
-	public static double computeZ(Matrix c, Matrix beta, double variance) {
-		return compute_cBetaHat(c, beta) / Math.sqrt(variance);
+	public static float computeZ(Matrix c, Matrix beta, float variance) {
+		return compute_cBetaHat(c, beta) / (float)Math.sqrt(variance);
 	}
 
-	public static double computeZ(double cBeta, double variance) {
+	public static float computeZ(float cBeta, float variance) {
 		return cBeta / variance;
 	}
 
@@ -40,7 +40,7 @@ public class Algorithm {
 	 * update 2019-04-12: Deprecated due to slowness, use the sparse version instead
 	 * var(c_beta_hat) = c(X'X)^(-1)X'D_r^(*)X(X'X)^(-1)c'
 	 */
-	public static double computeVariance(Matrix c, Matrix X, Matrix D) {
+	public static float computeVariance(Matrix c, Matrix X, Matrix D) {
 		try {
 			return c.mmul(X.tmmul(X).minv())
 					.mmult(X).mmul(D)
@@ -58,7 +58,7 @@ public class Algorithm {
 	 * computation
 	 * ~200 times faster than without using sparse BLAS routines.
 	 */
-	public static double computeVarianceSparse(Matrix c, Matrix X, Matrix D) {
+	public static float computeVarianceSparse(Matrix c, Matrix X, Matrix D) {
 		try {
 			return c.mmul(X.tmmul(X).minv()).mmult(X)
 					.mmul(D.smmul(X.mmul(X.tmmul(X).minv()))).mmult(c)
@@ -75,12 +75,12 @@ public class Algorithm {
 	 * computation
 	 * ~200 times faster than without using sparse BLAS routines.
 	 */
-	public static double compute_variance_sparse_fast(Matrix c, Matrix XTXInverseXT, Matrix XXTXInverse,
+	public static float compute_variance_sparse_fast(Matrix c, Matrix XTXInverseXT, Matrix XXTXInverse,
 			Matrix D) {
 		return c.mmul(XTXInverseXT).mmul(D.smmul(XXTXInverse)).mmult(c).get(0);
 	}
 
-	public static double compute_variance_sparse_fastest(Matrix c, Matrix XTXInverseXT, Matrix XXTXInverse,
+	public static float compute_variance_sparse_fastest(Matrix c, Matrix XTXInverseXT, Matrix XXTXInverse,
 			Matrix D) {
 		return c.mmul(XTXInverseXT).mmul(D.dmmul(XXTXInverse)).mmult(c).get(0);
 	}
@@ -94,29 +94,29 @@ public class Algorithm {
 	 * Formula: {(c*beta_hat-theta_0)'* Var(c*beta_hat)^-1 * (c*beta_hat-theta_0)
 	 * - (c*beta_hat-theta_1)'* Var(c*beta_hat)^-1 * (c*beta_hat-theta_1)} / 2
 	 */
-	public static double compute_SPRT(double cBetaHat, double thetaZero, double thetaOne, double variance) {
-		return (Math.pow((cBetaHat - thetaZero), 2) - Math.pow((cBetaHat - thetaOne), 2)) / (2 * variance);
+	public static float compute_SPRT(float cBetaHat, float thetaZero, float thetaOne, float variance) {
+		return (float)(Math.pow((cBetaHat - thetaZero), 2) - Math.pow((cBetaHat - thetaOne), 2)) / (2 * variance);
 	}
 
 	/**
-	 * Params: double -- alpha, double -- beta
+	 * Params: float -- alpha, float -- beta
 	 * Output: the stopping rule's boundaries [A, B]
 	 * Functionality: this method takes in two parameters, alpha and bate, and
 	 * computes the stopping rule's boundary values A and B
 	 */
-	public static double SPRTUpperBound(double alpha, double beta) {
-		double A = Math.log((1 - beta) / alpha);
-		double B = Math.log(beta / (1 - alpha));
+	public static float SPRTUpperBound(float alpha, float beta) {
+		float A = (float)Math.log((1 - beta) / alpha);
+		float B = (float)Math.log(beta / (1 - alpha));
 		return Math.max(A, B);
 	}
 
-	public static double SPRTLowerBound(double alpha, double beta) {
-		double A = Math.log((1 - beta) / alpha);
-		double B = Math.log(beta / (1 - alpha));
+	public static float SPRTLowerBound(float alpha, float beta) {
+		float A = (float)Math.log((1 - beta) / alpha);
+		float B = (float)Math.log(beta / (1 - alpha));
 		return Math.min(A, B);
 	}
 
-	public static int compute_activation_stat(double SPRT, double upper, double lower) {
+	public static int compute_activation_stat(float SPRT, float upper, float lower) {
 		if (SPRT > upper)
 			return 1;
 		else if (SPRT < lower)
@@ -125,17 +125,17 @@ public class Algorithm {
 			return 0;
 	}
 
-	public static double computeTheta1(double Z, double variance) {
-		return Z * Math.sqrt(variance);
+	public static float computeTheta1(float Z, float variance) {
+		return Z * (float)Math.sqrt(variance);
 	}
 
 	/*
 	 * compute H matrix and store its diagonal value into a vector
 	 * H = X * (X'X)^-1 * X'
 	 */
-	public static double[] computeH(Matrix XXTXInverse, Matrix X) {
+	public static float[] computeH(Matrix XXTXInverse, Matrix X) {
 		Matrix H = XXTXInverse.mmult(X);
-		double[] ret = new double[H.getRow()];
+		float[] ret = new float[H.getRow()];
 		for (int i = 0; i < H.getRow(); i++) {
 			ret[i] = H.get(i, i);
 		}
@@ -146,8 +146,8 @@ public class Algorithm {
 	 * r_i = Y_i - X * beta_hat
 	 * r_i is used to compute (D_r)^*
 	 */
-	public static double[] computeR(Matrix Y, Matrix X, Matrix beta) {
-		double[] ret = new double[Y.getRow()];
+	public static float[] computeR(Matrix Y, Matrix X, Matrix beta) {
+		float[] ret = new float[Y.getRow()];
 		Matrix XBeta = X.mmul(beta);
 		for (int i = 0; i < Y.getRow(); i++) {
 			ret[i] = Y.get(i, 0) - XBeta.get(i, 0);
@@ -159,11 +159,11 @@ public class Algorithm {
 	 * D_values is a single vector is used as a parameter of sparse matrix
 	 * computation
 	 */
-	public static Matrix generateD(double[] R, double[] H, MatrixStorageScope datatype) {
+	public static Matrix generateD(float[] R, float[] H, MatrixStorageScope datatype) {
 		Matrix D = new Matrix(R.length, R.length, datatype);
 		for (int i = 0; i < R.length; i++) {
 			for (int j = 0; j < R.length; j++) {
-				D.put(i * R.length + j,  Math.pow(R[i], 2) / (1 - H[i]));
+				D.put(i * R.length + j,  (float)(Math.pow(R[i], 2) / (1.0 - H[i])));
 			}
 		}
 		return D;
@@ -175,10 +175,10 @@ public class Algorithm {
 	 * r_i = Y_i - X * beta_hat
 	 */
 	@Deprecated
-	public static double estimateSigmaHatSquare(double[] response, Matrix X, Matrix beta, int scanNumber, int col) {
+	public static float estimateSigmaHatSquare(float[] response, Matrix X, Matrix beta, int scanNumber, int col) {
 		Matrix XBeta = X.mmul(beta);
-		double[] xBetaArray = XBeta.getArray();
-		double ret = 0.0;
+		float[] xBetaArray = XBeta.getArray();
+		float ret = 0.0f;
 		for (int i = 0; i < response.length; i++) {
 			ret += Math.pow((response[i] - xBetaArray[i]), 2);
 		}
@@ -187,11 +187,11 @@ public class Algorithm {
 	}
 
 	@Deprecated
-	public static double computeVarianceUsingSigmaHatSquare(double sigmaHatSquare, Matrix c, Matrix XTXInverse) {
+	public static float computeVarianceUsingSigmaHatSquare(float sigmaHatSquare, Matrix c, Matrix XTXInverse) {
 		return sigmaHatSquare * c.mmul(XTXInverse).mmult(c).get(0);
 	}
 
-	public static int evaluateConfidenceInterval(double cBeta, double variance, double CI, double theta) {
+	public static int evaluateConfidenceInterval(float cBeta, float variance, float CI, float theta) {
 		if (cBeta + CI * Math.sqrt(variance) < theta)
 			return -1;
 		else if (cBeta - CI * Math.sqrt(variance) > 0)
@@ -200,15 +200,15 @@ public class Algorithm {
 			return 0;
 	}
 
-	public static double[] computeMeanAndVariance(double[] array) {
-		double[] ret = new double[2];
-		double total = 0.0;
-		for (double d : array) {
+	public static float[] computeMeanAndVariance(float[] array) {
+		float[] ret = new float[2];
+		float total = 0.0f;
+		for (float d : array) {
 			total += d;
 		}
 		ret[0] = total / array.length;
-		double variance = 0.0;
-		for (double d : array) {
+		float variance = 0.0f;
+		for (float d : array) {
 			variance += Math.pow(d - ret[0], 2);
 		}
 		ret[1] = variance / array.length;
