+++ src/main/java/sprt/Config.java
@@ -13,15 +13,15 @@ public class Config implements Serializable {
 	public int MAX_SCAN = 238; // total scans in the session
 	public int COL = 8; // # of columns in design matrix / contrasts
 	public int K = 78; // first K scans used to collect data to estimate theta 1
-	public double ZScore = 3.12; // Z values used for estimating theta 1
-	public double theta0 = 0;
-	public double theta1 = 1;
-	public double alpha = 0.001; // used to decide SPRT boundary
-	public double beta = 0.1; // used to decide SPRT boundary
-	public double percentLower = 0.85; // percentage of SPRT to count as active
-	public double percentUpper = 0.85; // percentage of SPRT to count as non-active
-	public double SPRTUpperBound = Algorithm.SPRTUpperBound(alpha, beta);
-	public double SPRTLowerBound = Algorithm.SPRTLowerBound(alpha, beta);
+	public float ZScore = 3.12f; // Z values used for estimating theta 1
+	public float theta0 = 0f;
+	public float theta1 = 1f;
+	public float alpha = 0.001f; // used to decide SPRT boundary
+	public float beta = 0.1f; // used to decide SPRT boundary
+	public float percentLower = 0.85f; // percentage of SPRT to count as active
+	public float percentUpper = 0.85f; // percentage of SPRT to count as non-active
+	public float SPRTUpperBound = Algorithm.SPRTUpperBound(alpha, beta);
+	public float SPRTLowerBound = Algorithm.SPRTLowerBound(alpha, beta);
 	public String BOLDPath = "dat/";
 	public String BOLDPrefix = "bold";
 	public Contrasts contrasts;
